refactor(semantic): hoist find_callees_in_range envelope into common::run_callee_query [dirge-ftmo]
8 of the 10 language adapters (python/go/rust/java/c/cpp/ruby/typescript)
implemented find_callees_in_range with a byte-identical parser→query→
collect→sort/dedup envelope, differing only in the language + the query
string. Hoist that envelope into semantic::common::run_callee_query(lang,
query_str, source, range); each adapter collapses to its lang + query_str +
one delegate call (and sheds the now-unused Query/QueryCursor/find_node_at_range/
StreamingIterator imports).

clojure + elixir are deliberately LEFT bespoke — they do post-capture
filtering (SPECIAL_FORMS / DEF_KW) the generic envelope doesn't, so they're
genuinely not the same shape (the audit overcounted them as identical).

~230 LOC removed. All 10 find_callees tests pass; clean under -D warnings --all-features.
